Initial commit.
It's an old project of mine, so it's nearly code complete, I just have to add the CLI (and document the code, but that's not a code-complete issue). Admittedly, "feature complete" isn't saying a whole lot in this case - it doesn't even support character classes. I might go back and add features (read: make the project actually potentially useful) at some point, but not now.